refactor(db,mcp): hoist resolve_steps_to_heads into WorkflowRepository
Lifts the private `build_resolved_steps` function and `ResolvedStep` struct
out of epigraph-mcp into `WorkflowRepository::resolve_steps_to_heads` in the
db crate. MCP caller updated to delegate; JSON shape and all 6 step_versioning
tests preserved.
