+++ crates/epigraph-db/src/repos/workflow.rs
@@ -6,6 +6,9 @@
 use sqlx::PgPool;
 use uuid::Uuid;
 
+use crate::errors::DbError;
+use crate::repos::claim::{ClaimRepository, LineageHead};
+
 /// Workflow recall result (semantic or text search).
 #[derive(Debug, Clone, serde::Serialize)]
 pub struct WorkflowRecallResult {
@@ -51,6 +54,23 @@ pub struct HierarchicalWorkflowRow {
     pub parent_id: Option<Uuid>,
     pub metadata: serde_json::Value,
     pub created_at: chrono::DateTime<chrono::Utc>,
+}
+
+/// Per-step resolution result for `find_workflow_hierarchical(resolve_to_latest=true)`.
+///
+/// `frozen_claim_id` is the original step claim attached to the workflow via
+/// the `executes` edge. `heads` lists the current head(s) of the step's
+/// lineage (claims with `step_lineage_id = $lineage` and no incoming
+/// `supersedes` edge); empty when the step has no `step_lineage_id` (legacy)
+/// or when the lineage has been pruned. `pending_resolution` is true when
+/// the lineage has more than one head — caller must reconcile before reuse.
+#[derive(Debug, serde::Serialize)]
+pub struct ResolvedStep {
+    pub step_index: usize,
+    pub frozen_claim_id: Uuid,
+    pub step_lineage_id: Option<Uuid>,
+    pub heads: Vec<LineageHead>,
+    pub pending_resolution: bool,
 }
 
 pub struct WorkflowRepository;
@@ -310,6 +330,52 @@ impl WorkflowRepository {
         .await?;
 
         Ok(root.map(|(id,)| id).unwrap_or(workflow_id))
+    }
+
+    /// For each `executes`-edge from the workflow root, walk supersedes/revises
+    /// edges to the latest head claim. Mirrors the resolution logic that
+    /// previously lived in `epigraph-mcp::tools::workflow_hierarchical::build_resolved_steps`.
+    ///
+    /// # Errors
+    /// Returns `DbError` if the database query fails.
+    pub async fn resolve_steps_to_heads(
+        pool: &PgPool,
+        workflow_id: Uuid,
+    ) -> Result<Vec<ResolvedStep>, DbError> {
+        // Pull all level=2 step claims under this workflow with their
+        // step_lineage_id, ordered by edge created_at + claim id (matches
+        // do_report_hierarchical_outcome_via_pool).
+        let step_rows: Vec<(Uuid, Option<Uuid>)> = sqlx::query_as(
+            "SELECT c.id, c.step_lineage_id \
+             FROM edges e \
+             JOIN claims c ON c.id = e.target_id \
+             WHERE e.source_id = $1 AND e.relationship = 'executes' AND (c.properties->>'level')::int = 2 \
+             ORDER BY e.created_at ASC, c.id ASC",
+        )
+        .bind(workflow_id)
+        .fetch_all(pool)
+        .await
+        .map_err(DbError::from)?;
+
+        let mut resolved = Vec::with_capacity(step_rows.len());
+        for (step_index, (frozen_claim_id, step_lineage_id)) in step_rows.into_iter().enumerate() {
+            let heads: Vec<LineageHead> = if let Some(lineage_id) = step_lineage_id {
+                ClaimRepository::latest_in_lineage(pool, lineage_id)
+                    .await
+                    .map_err(DbError::from)?
+            } else {
+                Vec::new()
+            };
+            let pending_resolution = heads.len() > 1;
+            resolved.push(ResolvedStep {
+                step_index,
+                frozen_claim_id,
+                step_lineage_id,
+                heads,
+                pending_resolution,
+            });
+        }
+        Ok(resolved)
     }
 
     /// Search hierarchical workflows by free-text query against `goal` and
