+++ crates/epigraph-mcp/src/tools/workflow_hierarchical.rs
@@ -13,7 +13,6 @@
 //! `GET /api/v1/workflows/hierarchical/search` and
 //! `POST /api/v1/workflows/hierarchical/:id/outcome`.
 
-use epigraph_db::{ClaimRepository, LineageHead};
 use rmcp::model::*;
 use uuid::Uuid;
 
@@ -77,7 +76,9 @@ pub async fn find_workflow_hierarchical(
         });
 
         if resolve_to_latest {
-            let resolved = build_resolved_steps(&server.pool, r.id).await?;
+            let resolved = epigraph_db::WorkflowRepository::resolve_steps_to_heads(&server.pool, r.id)
+                .await
+                .map_err(internal_error)?;
             entry["resolved_steps"] = serde_json::to_value(resolved).map_err(internal_error)?;
         }
 
@@ -89,63 +90,6 @@ pub async fn find_workflow_hierarchical(
         "total": workflows.len(),
         "resolve_to_latest": resolve_to_latest,
     }))
-}
-
-/// Per-step resolution result for `find_workflow_hierarchical(resolve_to_latest=true)`.
-///
-/// `frozen_claim_id` is the original step claim attached to the workflow via
-/// the `executes` edge. `heads` lists the current head(s) of the step's
-/// lineage (claims with `step_lineage_id = $lineage` and no incoming
-/// `supersedes` edge); empty when the step has no `step_lineage_id` (legacy)
-/// or when the lineage has been pruned. `pending_resolution` is true when
-/// the lineage has more than one head — caller must reconcile before reuse.
-#[derive(serde::Serialize)]
-struct ResolvedStep {
-    step_index: usize,
-    frozen_claim_id: Uuid,
-    step_lineage_id: Option<Uuid>,
-    heads: Vec<LineageHead>,
-    pending_resolution: bool,
-}
-
-async fn build_resolved_steps(
-    pool: &sqlx::PgPool,
-    workflow_id: Uuid,
-) -> Result<Vec<ResolvedStep>, McpError> {
-    // Pull all level=2 step claims under this workflow with their
-    // step_lineage_id, ordered by edge created_at + claim id (matches
-    // do_report_hierarchical_outcome_via_pool).
-    let step_rows: Vec<(Uuid, Option<Uuid>)> = sqlx::query_as(
-        "SELECT c.id, c.step_lineage_id \
-         FROM edges e \
-         JOIN claims c ON c.id = e.target_id \
-         WHERE e.source_id = $1 AND e.relationship = 'executes' AND (c.properties->>'level')::int = 2 \
-         ORDER BY e.created_at ASC, c.id ASC",
-    )
-    .bind(workflow_id)
-    .fetch_all(pool)
-    .await
-    .map_err(internal_error)?;
-
-    let mut resolved = Vec::with_capacity(step_rows.len());
-    for (step_index, (frozen_claim_id, step_lineage_id)) in step_rows.into_iter().enumerate() {
-        let heads = if let Some(lineage_id) = step_lineage_id {
-            ClaimRepository::latest_in_lineage(pool, lineage_id)
-                .await
-                .map_err(internal_error)?
-        } else {
-            Vec::new()
-        };
-        let pending_resolution = heads.len() > 1;
-        resolved.push(ResolvedStep {
-            step_index,
-            frozen_claim_id,
-            step_lineage_id,
-            heads,
-            pending_resolution,
-        });
-    }
-    Ok(resolved)
 }
 
 // ── report_hierarchical_outcome ────────────────────────────────────────────
